# Improve PolygonLayer rendering performance