--- a/example/lib/pages/polygon.dart
+++ b/example/lib/pages/polygon.dart
@@ -6,49 +6,63 @@ import 'package:latlong2/latlong.dart';
 class PolygonPage extends StatelessWidget {
   static const String route = 'polygon';
 
-  const PolygonPage({Key? key}) : super(key: key);
+  PolygonPage({Key? key}) : super(key: key) {
+    const size = 0.5;
+    for (double x = 0; x < 50; x += size * 2) {
+      for (double y = 0; y < 30; y += size * 2) {
+        lotsOfPolys.add(<LatLng>[
+          LatLng(0 + x, 0 + y),
+          LatLng(0 + x, size + y),
+          LatLng(size + x, size + y),
+          LatLng(size + x, 0 + y),
+        ]);
+      }
+    }
+  }
+
+  late final lotsOfPolys = <List<LatLng>>[];
+
+  final notFilledPoints = <LatLng>[
+    LatLng(51.5, -0.09),
+    LatLng(53.3498, -6.2603),
+    LatLng(48.8566, 2.3522),
+  ];
+
+  final filledPoints = <LatLng>[
+    LatLng(55.5, -0.09),
+    LatLng(54.3498, -6.2603),
+    LatLng(52.8566, 2.3522),
+  ];
+
+  final notFilledDotedPoints = <LatLng>[
+    LatLng(49.29, -2.57),
+    LatLng(51.46, -6.43),
+    LatLng(49.86, -8.17),
+    LatLng(48.39, -3.49),
+  ];
+
+  final filledDotedPoints = <LatLng>[
+    LatLng(46.35, 4.94),
+    LatLng(46.22, -0.11),
+    LatLng(44.399, 1.76),
+  ];
+
+  final labelPoints = <LatLng>[
+    LatLng(60.16, -9.38),
+    LatLng(60.16, -4.16),
+    LatLng(61.18, -4.16),
+    LatLng(61.18, -9.38),
+  ];
+
+  final labelRotatedPoints = <LatLng>[
+    LatLng(58.21, -10.28),
+    LatLng(58.21, -7.01),
+    LatLng(59.77, -7.01),
+    LatLng(59.77, -10.28),
+  ];
 
   @override
   Widget build(BuildContext context) {
-    final notFilledPoints = <LatLng>[
-      LatLng(51.5, -0.09),
-      LatLng(53.3498, -6.2603),
-      LatLng(48.8566, 2.3522),
-    ];
-
-    final filledPoints = <LatLng>[
-      LatLng(55.5, -0.09),
-      LatLng(54.3498, -6.2603),
-      LatLng(52.8566, 2.3522),
-    ];
-
-    final notFilledDotedPoints = <LatLng>[
-      LatLng(49.29, -2.57),
-      LatLng(51.46, -6.43),
-      LatLng(49.86, -8.17),
-      LatLng(48.39, -3.49),
-    ];
-
-    final filledDotedPoints = <LatLng>[
-      LatLng(46.35, 4.94),
-      LatLng(46.22, -0.11),
-      LatLng(44.399, 1.76),
-    ];
-
-    final labelPoints = <LatLng>[
-      LatLng(60.16, -9.38),
-      LatLng(60.16, -4.16),
-      LatLng(61.18, -4.16),
-      LatLng(61.18, -9.38),
-    ];
-
-    final labelRotatedPoints = <LatLng>[
-      LatLng(58.21, -10.28),
-      LatLng(58.21, -7.01),
-      LatLng(59.77, -7.01),
-      LatLng(59.77, -10.28),
-    ];
-
     return Scaffold(
       appBar: AppBar(title: const Text('Polygons')),
       drawer: buildDrawer(context, PolygonPage.route),
@@ -73,6 +87,14 @@ class PolygonPage extends StatelessWidget {
                     userAgentPackageName: 'dev.fleaflet.flutter_map.example',
                   ),
                   PolygonLayer(polygons: [
+                    //Many polygons that generate the same renderHashCode sequentially
+                    //will batch for a large performance increase.
+                    for(final poly in lotsOfPolys)
+                      Polygon(
+                        points: poly,
+                        borderStrokeWidth: 2,
+                        borderColor: Colors.orange,
+                        ),
                     Polygon(
                       points: notFilledPoints,
                       isFilled: false, // By default it's false