--- a/lib/src/layer/polygon_layer.dart
+++ b/lib/src/layer/polygon_layer.dart
@@ -1,5 +1,7 @@
 import 'dart:math';
 
+import 'package:collection/collection.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/src/layer/label.dart';
@@ -13,9 +15,7 @@ enum PolygonLabelPlacement {
 
 class Polygon {
   final List<LatLng> points;
-  final List<Offset> offsets = [];
   final List<List<LatLng>>? holePointsList;
-  final List<List<Offset>>? holeOffsetsList;
   final Color color;
   final double borderStrokeWidth;
   final Color borderColor;
@@ -24,11 +24,13 @@ class Polygon {
   final bool isFilled;
   final StrokeCap strokeCap;
   final StrokeJoin strokeJoin;
-  late final LatLngBounds boundingBox;
   final String? label;
   final TextStyle labelStyle;
   final PolygonLabelPlacement labelPlacement;
   final bool rotateLabel;
+
+  /// Calculated value
+  final LatLngBounds boundingBox;
 
   Polygon({
     required this.points,
@@ -45,9 +47,11 @@ class Polygon {
     this.labelStyle = const TextStyle(),
     this.labelPlacement = PolygonLabelPlacement.centroid,
     this.rotateLabel = false,
-  }) : holeOffsetsList = null == holePointsList || holePointsList.isEmpty
-            ? null
-            : List.generate(holePointsList.length, (_) => []);
+  }) : boundingBox = LatLngBounds.fromPoints(points);
+
+  /// Used to batch draw calls to the canvas.
+  int get renderHashCode => Object.hash(color, borderStrokeWidth, borderColor,
+      isDotted, isFilled, strokeCap, strokeJoin);
 }
 
 class PolygonLayer extends StatelessWidget {
@@ -56,17 +60,11 @@ class PolygonLayer extends StatelessWidget {
   /// screen space culling of polygons based on bounding box
   final bool polygonCulling;
 
-  PolygonLayer({
+  const PolygonLayer({
     super.key,
     this.polygons = const [],
     this.polygonCulling = false,
-  }) {
-    if (polygonCulling) {
-      for (final polygon in polygons) {
-        polygon.boundingBox = LatLngBounds.fromPoints(polygon.points);
-      }
-    }
-  }
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -74,106 +72,178 @@ class PolygonLayer extends StatelessWidget {
       builder: (BuildContext context, BoxConstraints bc) {
         final map = FlutterMapState.maybeOf(context)!;
         final size = Size(bc.maxWidth, bc.maxHeight);
-        final polygonsWidget = <Widget>[];
 
-        for (final polygon in polygons) {
-          polygon.offsets.clear();
-
-          if (null != polygon.holeOffsetsList) {
-            for (final offsets in polygon.holeOffsetsList!) {
-              offsets.clear();
-            }
-          }
-
-          if (polygonCulling &&
-              !polygon.boundingBox.isOverlapping(map.bounds)) {
-            // skip this polygon as it's offscreen
-            continue;
-          }
-
-          _fillOffsets(polygon.offsets, polygon.points, map);
-
-          if (null != polygon.holePointsList) {
-            final len = polygon.holePointsList!.length;
-            for (var i = 0; i < len; ++i) {
-              _fillOffsets(
-                  polygon.holeOffsetsList![i], polygon.holePointsList![i], map);
-            }
-          }
-
-          polygonsWidget.add(
-            CustomPaint(
-              painter: PolygonPainter(polygon, map.rotationRad),
-              size: size,
-            ),
-          );
-        }
-
-        return Stack(
-          children: polygonsWidget,
+        return CustomPaint(
+          painter: PolygonPainter(
+              polygonOpts: polygons,
+              rotationRad: map.rotationRad,
+              map: map,
+              polygonCulling: polygonCulling),
+          size: size,
         );
       },
     );
   }
-
-  void _fillOffsets(final List<Offset> offsets, final List<LatLng> points,
-      FlutterMapState map) {
-    final len = points.length;
-    for (var i = 0; i < len; ++i) {
-      final point = points[i];
-      final offset = map.getOffsetFromOrigin(point);
-      offsets.add(offset);
-    }
-  }
 }
 
 class PolygonPainter extends CustomPainter {
-  final Polygon polygonOpt;
+  final List<Polygon> polygonOpts;
   final double rotationRad;
+  final FlutterMapState map;
+  final bool polygonCulling;
 
-  PolygonPainter(this.polygonOpt, this.rotationRad);
+  PolygonPainter(
+      {required this.polygonOpts,
+      required this.rotationRad,
+      required this.map,
+      required this.polygonCulling});
 
   @override
   void paint(Canvas canvas, Size size) {
-    if (polygonOpt.offsets.isEmpty) {
-      return;
-    }
     final rect = Offset.zero & size;
-    _paintPolygon(canvas, rect);
+
+    final batches = <List<Polygon>>[];
+    //Batch sequentially ordered polygons with the same rendering information
+    int? lastHash;
+    for (final polygon in polygonOpts) {
+      if (polygonCulling && !polygon.boundingBox.isOverlapping(map.bounds)) {
+        // skip this polygon as it's offscreen
+        continue;
+      }
+      final hash = polygon.renderHashCode;
+      if (hash != lastHash) {
+        batches.add([polygon]);
+        lastHash = hash;
+      } else {
+        batches.last.add(polygon);
+      }
+    }
+
+    print("batch_rate ${1-(batches.length/polygonOpts.length)}");
+
+    for (final batch in batches) {
+      final fillPaint = Paint();
+      final fillPath = Path();
+
+      final borderPaint = Paint();
+      final borderPath = Path();
+
+      fillPaint
+        ..style =
+            batch.first.isFilled ? PaintingStyle.fill : PaintingStyle.stroke
+        ..color = batch.first.color;
+
+      borderPaint
+        ..color = batch.first.borderColor
+        ..strokeWidth = batch.first.borderStrokeWidth
+        ..style = PaintingStyle.stroke
+        ..strokeCap = batch.first.strokeCap
+        ..strokeJoin = batch.first.strokeJoin;
+
+      //Render each polygon on the canvas.
+      for (final polygon in batch) {
+        //Offsets are potentially able to be pre-computed at a given projection
+        //then transformed based on zoom + center.
+        final polygonOffsets = List.generate(polygon.points.length,
+            (index) => map.getOffsetFromOrigin(polygon.points[index]),
+            growable: false);
+
+        if (polygon.holePointsList == null) {
+          //Polygon without holes
+          fillPath.addPolygon(polygonOffsets, true);
+
+          _paintBorder(canvas, polygon, polygonOffsets, null, borderPaint, borderPath);
+        } else {
+          //Polygon with holes
+          final holePolygonPaint = Paint()
+            ..style =
+                batch.first.isFilled ? PaintingStyle.fill : PaintingStyle.stroke
+            ..color = batch.first.color;
+
+          //TODO this is kinda an ugly one-liner.
+          final polygonHoleOffsets = List.generate(
+              polygon.holePointsList!.length,
+              (poly) => List.generate(
+                  polygon.holePointsList![poly].length,
+                  (index) => map.getOffsetFromOrigin(
+                      polygon.holePointsList![poly][index]),
+                  growable: false),
+              growable: false);
+
+          canvas.saveLayer(rect, holePolygonPaint);
+          holePolygonPaint.style = PaintingStyle.fill;
+
+          for (final offsets in polygonHoleOffsets) {
+            final path = Path();
+            path.addPolygon(offsets, true);
+            canvas.drawPath(path, holePolygonPaint);
+          }
+
+          holePolygonPaint
+            ..color = polygon.color
+            ..blendMode = BlendMode.srcOut;
+
+          final path = Path();
+          fillPath.addPolygon(polygonOffsets, true);
+          canvas.drawPath(path, holePolygonPaint);
+
+          _paintBorder(
+              canvas, polygon, polygonOffsets, polygonHoleOffsets, borderPaint, borderPath);
+
+          canvas.restore();
+        }
+      }
+
+      //Draw the polygon fills
+      canvas.drawPath(fillPath, fillPaint);
+      //Draw the polygon border
+      canvas.drawPath(borderPath, borderPaint);
+
+      //Draw labels
+      for (final polygon in batch) {
+        if (polygon.label != null) {
+          //It's probably fine to do this since the calculation is fairly lightweight.
+          final polygonOffsets = List.generate(polygon.points.length,
+              (index) => map.getOffsetFromOrigin(polygon.points[index]),
+              growable: false);
+          Label.paintText(
+            canvas,
+            polygonOffsets,
+            polygon.label,
+            polygon.labelStyle,
+            rotationRad,
+            rotate: polygon.rotateLabel,
+            labelPlacement: polygon.labelPlacement,
+          );
+        }
+      }
+    }
+    canvas.clipRect(rect);
   }
 
-  void _paintBorder(Canvas canvas) {
-    if (polygonOpt.borderStrokeWidth > 0.0) {
-      final borderPaint = Paint()
-        ..color = polygonOpt.borderColor
-        ..strokeWidth = polygonOpt.borderStrokeWidth;
+  void _paintBorder(Canvas canvas, Polygon polygon, List<Offset> polygonOffsets,
+      List<List<Offset>>? polygonHoleOffsets, Paint borderPaint, Path solidPath) {
+    if (polygon.borderStrokeWidth > 0.0) {
+      if (polygon.isDotted) {
+        // Dotted polygon
+        final borderRadius = (polygon.borderStrokeWidth / 2);
 
-      if (polygonOpt.isDotted) {
-        final borderRadius = (polygonOpt.borderStrokeWidth / 2);
-
-        final spacing = polygonOpt.borderStrokeWidth * 1.5;
+        final spacing = polygon.borderStrokeWidth * 1.5;
         _paintDottedLine(
-            canvas, polygonOpt.offsets, borderRadius, spacing, borderPaint);
+            canvas, polygonOffsets, borderRadius, spacing, borderPaint);
 
-        if (!polygonOpt.disableHolesBorder &&
-            null != polygonOpt.holeOffsetsList) {
-          for (final offsets in polygonOpt.holeOffsetsList!) {
+        if (!polygon.disableHolesBorder && null != polygonHoleOffsets) {
+          for (final offsets in polygonHoleOffsets) {
             _paintDottedLine(
                 canvas, offsets, borderRadius, spacing, borderPaint);
           }
         }
       } else {
-        borderPaint
-          ..style = PaintingStyle.stroke
-          ..strokeCap = polygonOpt.strokeCap
-          ..strokeJoin = polygonOpt.strokeJoin;
-
-        _paintLine(canvas, polygonOpt.offsets, borderPaint);
-
-        if (!polygonOpt.disableHolesBorder &&
-            null != polygonOpt.holeOffsetsList) {
-          for (final offsets in polygonOpt.holeOffsetsList!) {
-            _paintLine(canvas, offsets, borderPaint);
+        //Solid polygon
+        solidPath.addPolygon(polygonOffsets, true);
+        if (!polygon.disableHolesBorder && null != polygonHoleOffsets) {
+          for (final offsets in polygonHoleOffsets) {
+            solidPath.addPolygon(offsets, true);
           }
         }
       }
@@ -202,77 +272,12 @@ class PolygonPainter extends CustomPainter {
     canvas.drawCircle(offsets.last, radius, paint);
   }
 
-  void _paintLine(Canvas canvas, List<Offset> offsets, Paint paint) {
-    if (offsets.isEmpty) {
-      return;
-    }
-    final path = Path()..addPolygon(offsets, true);
-    canvas.drawPath(path, paint);
-  }
-
-  void _paintPolygon(Canvas canvas, Rect rect) {
-    final paint = Paint();
-
-    if (null != polygonOpt.holeOffsetsList) {
-      canvas.saveLayer(rect, paint);
-      paint.style = PaintingStyle.fill;
-
-      for (final offsets in polygonOpt.holeOffsetsList!) {
-        final path = Path();
-        path.addPolygon(offsets, true);
-        canvas.drawPath(path, paint);
-      }
-
-      paint
-        ..color = polygonOpt.color
-        ..blendMode = BlendMode.srcOut;
-
-      final path = Path();
-      path.addPolygon(polygonOpt.offsets, true);
-      canvas.drawPath(path, paint);
-
-      _paintBorder(canvas);
-
-      canvas.restore();
-    } else {
-      canvas.clipRect(rect);
-      paint
-        ..style =
-            polygonOpt.isFilled ? PaintingStyle.fill : PaintingStyle.stroke
-        ..color = polygonOpt.color;
-
-      final path = Path();
-      path.addPolygon(polygonOpt.offsets, true);
-      canvas.drawPath(path, paint);
-
-      _paintBorder(canvas);
-
-      if (polygonOpt.label != null) {
-        Label.paintText(
-          canvas,
-          polygonOpt.offsets,
-          polygonOpt.label,
-          polygonOpt.labelStyle,
-          rotationRad,
-          rotate: polygonOpt.rotateLabel,
-          labelPlacement: polygonOpt.labelPlacement,
-        );
-      }
-    }
-  }
-
   @override
   bool shouldRepaint(PolygonPainter oldDelegate) => false;
 
-  double _dist(Offset v, Offset w) {
-    return sqrt(_dist2(v, w));
-  }
+  double _dist(Offset v, Offset w) => sqrt(_dist2(v, w));
 
-  double _dist2(Offset v, Offset w) {
-    return _sqr(v.dx - w.dx) + _sqr(v.dy - w.dy);
-  }
+  double _dist2(Offset v, Offset w) => _sqr(v.dx - w.dx) + _sqr(v.dy - w.dy);
 
-  double _sqr(double x) {
-    return x * x;
-  }
+  double _sqr(double x) => x * x;
 }