Changes:
- Better Offset generation code. Uses fixed-sized lists now.
- Added batching code. Every sequential polygon that generates the same renderHashCode will have optimized draw calls to the canvas. In many cases this can more than double the performance. With 15 batches and 1.5k polygons I profiled a framerate change from 13 fps -> 38 fps. 
- Batching code overhead is minimal. I profiled <1ms on a high end phone (z fold 4), and <2ms on a budget phone (pixel 4a) with 1.5k polygons. I don't think there would be any situation where performance would become worse, combined with the other optimizations. (this needs to be tested though)
- Added batched polygons to example app.
- Polygons with holes now render labels (not sure if this is intentional to be the case or not).
- Refactored the renderer to de-duplicate the code a reasonable amount.


Possible side effects:
- Since polygons are batched there may potentially be slight differences between this renderer and the old one.
- Polygons with holes didn't get refactored because I have limited knowledge of how the hole renderer works (TODO),