@mohammedX6 2022-10-09
But I have a question, why are you not saving the offsets into the polygon offsets variable? you are doing the calculation of the offsets twice.