@MooNag 2022-10-15
I did this since the offsets are computed on every build anyways, and the compute function has low overhead. Potentially this could be optimized. Previously, the offsets were being stored in a growable list, which wasted multiple milliseconds of time just resizing lists, whereas now using a single ListGenerator the whole compute can be done in less time.

The offsets being calculated twice is only for if you are using a label. Since those are computed after the main polygons and I don't want to commit on having a cache variable for only a **tiny** amount of performance. Especially given that at this point the limiting factor actually seems to be the GPU rather than the CPU on most phones where many polygons are involved. Let me know if you have any questions! I can also post a profiler if you have any questions about the compute time percentages for each kind of thing.