@mohammedX6 2022-10-15
i learned new thing toady from you today!.
I already implemented your modification and i saw a huge amount of performance increase in my app especially on older tablets, you did a great job!