@JaffaKetchup 2022-11-15
Hey @MooNag. Did you ever fix the issue with the hole points not working? I think I reported to you on Discord. That needs to be fixed before we can merge.