@ignatz 2022-12-27
I also ran into performance issues with polygon/polylines. Instead of batching canvas paints, I experimented with minimizing repaints: https://github.com/fleaflet/flutter_map/pull/1423 (currently only for polylines) but the same approach should work for polygons, too. I'm mostly raising this here because I don't think the approaches would be compatible. Concretely: batching and culling require you to redraw on every pan (unless you also merge bounding boxes).  It might be worth thinking a little bit about with approach might be more promising going forward, i.e. minimizing number of canvases vs minimizing number of draws.